+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renamed `TuningResourceType.Street` to `TuningResourceType.Tuning`.
 ### Added
 - Added various types/interfaces to API in `@s4tk/models/types`.
+- Added `cloneKeys` and `cloneResources` options to `Package.create()` options.
 
 ## [0.1.1] - 2022/01/29
 ### Added